Renaming and extra comments
